requirements.txt
Flask
openai
pyTelegramBotApi
googletrans==3.1.0a0
python-dotenv
gunicorn
psycopg2-binary